agent/requirements.txt
anthropic>=0.49.0
flask==3.0.3
flask-cors==4.0.0
python-nmap==0.7.1
scapy==2.5.0
mysql-connector-python==8.3.0
python-dotenv==1.0.1
requests==2.31.0
langgraph==0.2.28
langchain-core==0.3.15
mcp==1.6.0
reportlab==4.2.0
